Applied patch from Borut Razem (mailto:borut.razem@siol.net) that enables gpsim to be compiled under windows using Visual Studio.

+++ src/breakpoints.cc
@@ -35,7 +35,7 @@ extern "C"{
 }
 
 extern int last_command_is_repeatable;
-extern unsigned int simulation_start_cycle;
+extern guint64 simulation_start_cycle;
 extern pic_processor *active_cpu;
 //extern "C" {
   extern void redisplay_prompt(void);  // in input.cc
